fix(ENTESB-11945): Move Json schema utils to separate package
