--- app/common/util/src/main/java/io/syndesis/common/util/json/JsonSchemaMixIn.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2016 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.syndesis.common.util.json;
-
-import com.fasterxml.jackson.annotation.JsonAlias;
-
-/**
- * @author Christoph Deppisch
- */
-public interface JsonSchemaMixIn {
-
-    /**
-     * Adds compatibility for draft-06 specific properties.
-     */
-    interface Draft6 {
-        @JsonAlias("$id") String getId();
-    }
-}
